Renamed test suite and moved to appropriate location
